+++ test/reservoir.js
@@ -1,0 +1,9 @@
+
+var reservoir = require('../libs/reservoir');
+
+describe('reservoir', function () {
+    it('should return reservoirData', function (done) {
+      	reservoir(done);
+    });
+});
+
